Remove supervisor upgrade information
This information isn't correct in the case that the device has
been hup'd. In these cases the supervisor will be reverted 
eventually and this can potentially cause data loss. We're
currently working on self-serve supervisor updates, but 
until then I think the safest route forward is to remove this
information.

Change-type: patch
Signed-off-by: Cameron Diver <cameron@balena.io>